Extract `CspDisableable` controller concern and add to a controller
This fixes a CSP problem that would occur if the user:
1. went to the add new quiz questions page
2. refreshed the page
3. submitted new question(s) (which goes back to the quiz show page)
