+++ app/controllers/quizzes_controller.rb
@@ -2,16 +2,11 @@
 
 class QuizzesController < ApplicationController
   extend Memoist
+  include CspDisableable
 
   self.container_classes = %w[py1 px3]
 
   before_action :set_quiz, only: %i[respondents leaderboard progress]
-
-  # workarounds until https://github.com/hotwired/turbo/issues/ 294 is resolved
-  before_action :dont_set_csp_nonce
-  content_security_policy do |policy|
-    policy.script_src(:self, :https, :unsafe_inline)
-  end
 
   def index
     authorize(Quiz, :index?)
@@ -63,10 +58,6 @@ class QuizzesController < ApplicationController
   end
 
   private
-
-  def dont_set_csp_nonce
-    request.env['action_dispatch.content_security_policy_nonce_directives'] = []
-  end
 
   def set_quiz
     quiz =
